The Code
The ipynb file is the code. You need to have python installed on your device, preferably jupyter notebook. Save the Ebola.txt file on your device and copy paste its path on your pc to the chdir(" path ") command in the code. You can also change the hyper parameters to try out stuff. 

If you want to create your own organism eg. Cat; just download cat genome as FASTA file from here :
https://www.ncbi.nlm.nih.gov/genome?term=felis%20catus 

provide the path and name of the txt file in the first cell of the code and you are good to go!